tests/inputdevice-mouse.html

<!DOCTYPE html>
<title>InputDevice test for mouse-triggered events</title>
<meta name="viewport" content="width=device-width">
<script src="https://w3c-test.org/resources/testharness.js"></script>
<script src="https://w3c-test.org/resources/testharnessreport.js"></script>
<script src="inputdevice-tests.js"></script>
<style>
#target {
  border: 1px solid black;
  width: 100px;
  height: 100px;
  background-color: lightgreen;
  margin-bottom: 20px;
}
</style>

<h1>InputDevice - test for mouse-triggered events</h1>
<p>Double-click with a pointing device that doesn't fire touch events (eg. a mouse) in the box below, then click outside the box.</p>
<div id="target" tabindex=1></div>
<div><b>Expected events:</b> <span id="eventList"></span></div>
<div id="log"></div>
<script>
'use strict';

function inputDeviceValidator(sourceDevice) {
  assert_true(event.sourceDevice != null, "sourceDevice is set");
  assert_false(event.sourceDevice.firesTouchEvents, "sourcedevice.firesTouchEvents");
}

function runTests() {
  var events = ["mousedown", "mousemove", "mouseup", "mouseenter", "mouseleave", "mouseover", "mouseout", "click", "dblclick", "focus", "focusin", "focusout", "blur"];
  if ('PointerEvent' in window)
    events = events.join(["pointerdown", "pointermove", "pointerup", "pointerenter", "pointerleave", "pointerover", "pointerout"]);
  runInputDeviceEventTests(events, inputDeviceValidator);
}
</script>
